Avoid misleading naming
It's misleading to refer to a `.cabal` being a project rather than a package, since there's also the `cabal.project` file to denote an actual project comprised of one or more packages as well as project-specific settings. `.cabal` files are specifications of the unit of distributions, they don't describe projects.